requirements.txt
Django
djangorestframework
django-recurrence

django-cors-headers
pillow
djangorestframework-simplejwt
